[BENCH-184] Fixup the build caching
Don't attempt to cache a build artifact if we don't have a valid URL
for the build artifact.   This can happen if we try to get the build
before the artifacts have been populated in bamboo.  If we don't have
the build artifact when we go to get it (get_build) then try to cache
it again.